safari-534.53-branch/LayoutTests/accessibility/aria-readonly-expected.txt

     
This tests that the aria-readonly attribute works. The first and third text fields should not be writable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS succeeded is false
PASS succeeded is true
PASS succeeded is false
PASS succeeded is true
PASS succeeded is false
PASS succeeded is true
PASS successfullyParsed is true

TEST COMPLETE
